pyproject.toml
[tool.poetry]
name = "ledgerflow"
version = "0.1.0"
description = "An Event-Sourced Ledger architecture using Kafka."
authors = ["Your Name <you@example.com>"]
packages = [
    { include = "command_api" },
    { include = "processor" },
    { include = "projector" },
    { include = "query_api" },
    { include = "shared" },
]


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
pydantic = "^2.4.2"
sqlalchemy = "^2.0.21"
alembic = "^1.12.0"
aiokafka = "0.7.2"
kafka-python = "2.0.2"
psycopg2-binary = "^2.9.7"


[tool.poetry.dev-dependencies]
pytest = "^7.4.2"
httpx = "^0.25.0"
ruff = "^0.0.292"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[tool.ruff]
line-length = 100
extend-select = ["E", "F"]